Fixing broken support for `ARCHIVES_URL` in tuxlite-tbs theme.
This commit fixes an issue with users supplying their own custom `ARCHIVES_URL`
setting in their settings file while using the tuxlite-tbs theme.

The issue is that the `base.html` template didn't take the `ARCHIVES_URL`
variable into account, and therefore ignores is. This causes problems when users
want a custom archives URL.

This also fixes another part of the issue in the `archives.html` template, in
which article URLs were relative (not absolute) which causes breakage when
linking to articles.

This commit fixes all the above problems.